refactor: move AvailabilityFilter to its own file
